+++ CHANGELOG.md
@@ -9,4 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Added usage display if `help` option provided or no environment directories
+  provided
+- Added `upgrade` option to upgrade julia in place if exists before creating
+- Added `clear` option to clear environment directory if exists before creating
+- Added ability to create multiple environments at once
 - `venv` function to create self-contained julia environments